% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CthulhuRoller.R
\name{CthulhuRoller}
\alias{CthulhuRoller}
\title{AddModifier}
\description{
AddModifier

AddModifier
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CthulhuRoller$new()}}
\item \href{#method-Roll}{\code{CthulhuRoller$Roll()}}
\item \href{#method-AddModifier}{\code{CthulhuRoller$AddModifier()}}
\item \href{#method-MaxHardSuccess}{\code{CthulhuRoller$MaxHardSuccess()}}
\item \href{#method-MaxExtremeSuccess}{\code{CthulhuRoller$MaxExtremeSuccess()}}
\item \href{#method-MaxBotch}{\code{CthulhuRoller$MaxBotch()}}
\item \href{#method-clone}{\code{CthulhuRoller$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CthulhuRoller$new(dieSides = NA, label = NA, modDieSides = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Roll"></a>}}
\if{latex}{\out{\hypertarget{method-Roll}{}}}
\subsection{Method \code{Roll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CthulhuRoller$Roll()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddModifier"></a>}}
\if{latex}{\out{\hypertarget{method-AddModifier}{}}}
\subsection{Method \code{AddModifier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CthulhuRoller$AddModifier(Roll, Modifier)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Roll}}{The current roll that shall be modified.}

\item{\code{Modifier}}{A value of enum `CthulhuRoller$ModifyType`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The modified roll (value between 1 and the number 
of sides of this roller die).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaxHardSuccess"></a>}}
\if{latex}{\out{\hypertarget{method-MaxHardSuccess}{}}}
\subsection{Method \code{MaxHardSuccess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CthulhuRoller$MaxHardSuccess(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaxExtremeSuccess"></a>}}
\if{latex}{\out{\hypertarget{method-MaxExtremeSuccess}{}}}
\subsection{Method \code{MaxExtremeSuccess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CthulhuRoller$MaxExtremeSuccess(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaxBotch"></a>}}
\if{latex}{\out{\hypertarget{method-MaxBotch}{}}}
\subsection{Method \code{MaxBotch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CthulhuRoller$MaxBotch(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CthulhuRoller$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
