% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollLogger.R
\name{RollLogger}
\alias{RollLogger}
\title{An abstract R6 class representing rolls with a 
die of a specified number of sides.}
\description{
A roller has a number of die sides and may support modify rolls.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RollLogger$new()}}
\item \href{#method-Log}{\code{RollLogger$Log()}}
\item \href{#method-AsHtml}{\code{RollLogger$AsHtml()}}
\item \href{#method-print}{\code{RollLogger$print()}}
\item \href{#method-clone}{\code{RollLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RollLogger$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dieSides}}{Dies sides}

\item{\code{label}}{A label that could qualify as button label
to initiate this roller.}

\item{\code{modDieSides}}{Sides of a modifier die.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`invisible(self)`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Log"></a>}}
\if{latex}{\out{\hypertarget{method-Log}{}}}
\subsection{Method \code{Log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RollLogger$Log(ToLog)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsHtml"></a>}}
\if{latex}{\out{\hypertarget{method-AsHtml}{}}}
\subsection{Method \code{AsHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RollLogger$AsHtml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RollLogger$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RollLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
